web/requirements.txt
flask
requests
gunicorn
azure-monitor-opentelemetry
opentelemetry-sdk